add react dev dashboard
ruse dashboard
