+++ private/dev/index.html
@@ -42,6 +42,9 @@
     <main class="foo">
       <h1>Dashboard is here</h1>
 
+      <p><a href="/react.html">React dashboard</a></p>
+      <p><a href="/dragdrop.html">DragDrop</a></p>
+
       <!-- some inputs in a form to check focus management in Dashboard -->
       <form id="upload-form" action="/">
         <button type="button" id="pick-files">Pick Files</button><br />
