+++ private/dev/react.html
@@ -1,0 +1,14 @@
+<!doctype html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>React</title>
+    <link rel="icon" type="image/png" href="https://uppy.io/img/logo.svg" />
+  </head>
+
+  <body>
+    <div id="root"></div>
+    <script type="module" src="/React.jsx"></script>
+  </body>
+</html>
